move slicing code to separate file
